app/requirements.txt
Flask==3.0.0
joblib==1.3.2
mysql-connector-python==8.2.0
scikit-learn==1.3.2